+++ plugins/node/instrumentation-dataloader/test/dataloader.test.ts
@@ -140,6 +140,29 @@ describe('DataloaderInstrumentation', () => {
         message: 'Error message',
       });
     });
+
+    it('correctly uses dataloader name (if available)', async () => {
+      const namedDataloader = new Dataloader(
+        async keys => keys.map((_, idx) => idx),
+        { name: 'test-name' }
+      );
+
+      assert.strictEqual(await namedDataloader.load('test'), 0);
+
+      // We should have exactly two spans (one for .load and one for the following batch)
+      assert.strictEqual(memoryExporter.getFinishedSpans().length, 2);
+      const [batchSpan, loadSpan] = memoryExporter.getFinishedSpans();
+
+      if ((namedDataloader as { name?: string | null }).name === undefined) {
+        // For versions of dataloader package that does not support name, we should
+        // not be adding anything to the names
+        assert.strictEqual(loadSpan.name, 'dataloader.load');
+        assert.strictEqual(batchSpan.name, 'dataloader.batch');
+      } else {
+        assert.strictEqual(loadSpan.name, 'dataloader.load test-name');
+        assert.strictEqual(batchSpan.name, 'dataloader.batch test-name');
+      }
+    });
   });
 
   describe('loadMany', () => {
@@ -234,6 +257,33 @@ describe('DataloaderInstrumentation', () => {
       assert.deepStrictEqual(loadManySpan.status, {
         code: SpanStatusCode.UNSET,
       });
+    });
+
+    it('correctly uses a generated name in spans', async () => {
+      const namedDataloader = new Dataloader(
+        async keys => keys.map((_, idx) => idx),
+        { name: 'test-name' }
+      );
+
+      assert.deepStrictEqual(await namedDataloader.loadMany(['test']), [0]);
+
+      // We should have exactly three spans (one for .loadMany, one for the underlying .load
+      // and one for the following batch)
+      assert.strictEqual(memoryExporter.getFinishedSpans().length, 3);
+      const [batchSpan, loadSpan, loadManySpan] =
+        memoryExporter.getFinishedSpans();
+
+      if ((namedDataloader as { name?: string | null }).name === undefined) {
+        // For versions of dataloader package that does not support name, we should
+        // not be adding anything to the names
+        assert.strictEqual(batchSpan.name, 'dataloader.batch');
+        assert.strictEqual(loadManySpan.name, 'dataloader.loadMany');
+        assert.strictEqual(loadSpan.name, 'dataloader.load');
+      } else {
+        assert.strictEqual(batchSpan.name, 'dataloader.batch test-name');
+        assert.strictEqual(loadManySpan.name, 'dataloader.loadMany test-name');
+        assert.strictEqual(loadSpan.name, 'dataloader.load test-name');
+      }
     });
   });