+++ plugins/node/opentelemetry-instrumentation-aws-lambda/src/instrumentation.ts
@@ -15,6 +15,7 @@
  */
 
 import * as path from 'path';
+import * as fs from 'fs';
 
 import {
   InstrumentationBase,
@@ -115,8 +116,15 @@ export class AwsLambdaInstrumentation extends InstrumentationBase {
     // Lambda loads user function using an absolute path.
     let filename = path.resolve(taskRoot, moduleRoot, module);
     if (!filename.endsWith('.js')) {
-      // Patching infrastructure currently requires a filename when requiring with an absolute path.
-      filename += '.js';
+      // its impossible to know in advance if the user has a cjs or js file.
+      // check that the .js file exists otherwise fallback to next known possibility
+      try {
+        fs.statSync(`${filename}.js`);
+        filename += '.js';
+      } catch (e) {
+        // fallback to .cjs
+        filename += '.cjs';
+      }
     }
 
     return [
@@ -140,7 +148,7 @@ export class AwsLambdaInstrumentation extends InstrumentationBase {
               return moduleExports;
             },
             (moduleExports?: LambdaModule) => {
-              if (moduleExports == undefined) return;
+              if (moduleExports == null) return;
               diag.debug('Removing patch for lambda handler');
               this._unwrap(moduleExports, functionName);
             }