+++ plugins/node/opentelemetry-instrumentation-pg/src/utils.ts
@@ -34,7 +34,6 @@ import {
 import {
   PgClientExtended,
   PostgresCallback,
-  PgErrorCallback,
   PgPoolCallback,
   PgPoolExtended,
   PgParsedConnectionParams,
@@ -43,10 +42,6 @@ import { PgInstrumentationConfig } from './types';
 import type * as pgTypes from 'pg';
 import { PgInstrumentation } from './';
 import { safeExecuteInTheMiddle } from '@opentelemetry/instrumentation';
-
-function arrayStringifyHelper(arr: Array<unknown>): string {
-  return '[' + arr.toString() + ']';
-}
 
 /**
  * Helper function to get a low cardinality span name from whatever info we have
@@ -157,10 +152,26 @@ export function handleConfigQuery(
     instrumentationConfig.enhancedDatabaseReporting &&
     Array.isArray(queryConfig.values)
   ) {
-    span.setAttribute(
-      AttributeNames.PG_VALUES,
-      arrayStringifyHelper(queryConfig.values)
-    );
+    try {
+      const convertedValues = queryConfig.values.map(value => {
+        if (value == null) {
+          return 'null';
+        } else if (value instanceof Buffer) {
+          return value.toString();
+        } else if (typeof value === 'object') {
+          if (typeof value.toPostgres === 'function') {
+            return value.toPostgres();
+          }
+          return JSON.stringify(value);
+        } else {
+          //string, number
+          return value.toString();
+        }
+      });
+      span.setAttribute(AttributeNames.PG_VALUES, convertedValues);
+    } catch (e) {
+      diag.error('failed to stringify ', queryConfig.values, e);
+    }
   }
 
   // Set plan name attribute, if present
@@ -239,10 +250,7 @@ export function patchCallbackPGPool(
   };
 }
 
-export function patchClientConnectCallback(
-  span: Span,
-  cb: PgErrorCallback
-): PgErrorCallback {
+export function patchClientConnectCallback(span: Span, cb: Function): Function {
   return function patchedClientConnectCallback(
     this: pgTypes.Client,
     err: Error
@@ -254,7 +262,7 @@ export function patchClientConnectCallback(
       });
     }
     span.end();
-    cb.call(this, err);
+    cb.apply(this, arguments);
   };
 }