+++ plugins/node/opentelemetry-instrumentation-pg/test/pg.test.ts
@@ -45,11 +45,7 @@ import {
   SemanticAttributes,
   DbSystemValues,
 } from '@opentelemetry/semantic-conventions';
-import { isSupported } from './utils';
 import { addSqlCommenterComment } from '../src/utils';
-
-const pgVersion = require('pg/package.json').version;
-const nodeVersion = process.versions.node;
 
 const memoryExporter = new InMemorySpanExporter();
 
@@ -120,8 +116,6 @@ describe('pg', () => {
   }
 
   before(async function () {
-    const skipForUnsupported =
-      process.env.IN_TAV && !isSupported(nodeVersion, pgVersion);
     const skip = () => {
       // this.skip() workaround
       // https://github.com/mochajs/mocha/issues/2683#issuecomment-375629901
@@ -129,12 +123,6 @@ describe('pg', () => {
       this.skip();
     };
 
-    if (skipForUnsupported) {
-      console.error(
-        `  pg - skipped - node@${nodeVersion} and pg@${pgVersion} are not compatible`
-      );
-      skip();
-    }
     if (!shouldTest) {
       skip();
     }
@@ -257,6 +245,17 @@ describe('pg', () => {
         done();
       });
       assert.strictEqual(res, undefined, 'No promise is returned');
+    });
+
+    it('should pass the client connection object in the callback function', done => {
+      connClient.connect(function (err: Error) {
+        // Even though the documented signature for connect() callback is `(err) => void`
+        // `pg` actually also passes the client if the connection was successful and some
+        // packages(`knex`) might rely on that
+        // https://github.com/brianc/node-postgres/blob/master/packages/pg/lib/client.js#L282
+        assert.strictEqual(arguments[1], connClient);
+        done();
+      });
     });
 
     it('should return a promise if callback is not provided', done => {
@@ -511,6 +510,60 @@ describe('pg', () => {
         } catch (e: any) {
           assert.ok(false, e.message);
         }
+      });
+    });
+
+    describe('Check configuration enhancedDatabaseReporting:true', () => {
+      const obj = { type: 'Fiat', model: '500', color: 'white' };
+      const buf = Buffer.from('abc');
+      const objWithToPostgres = {
+        toPostgres: () => {
+          return 'custom value';
+        },
+      };
+      const query =
+        'SELECT $1::text as msg1, $2::bytea as bufferParam, $3::integer as numberParam, $4::jsonb as objectParam, $5::text as objToPostgres, $6::text as msg2, $7::text as msg3';
+      const values = [
+        'Hello,World',
+        buf,
+        6,
+        obj,
+        objWithToPostgres,
+        null,
+        undefined,
+      ];
+
+      const events: TimedEvent[] = [];
+
+      const attributes = {
+        ...DEFAULT_ATTRIBUTES,
+        [SemanticAttributes.DB_STATEMENT]: query,
+        [AttributeNames.PG_VALUES]: [
+          'Hello,World',
+          'abc',
+          '6',
+          '{"type":"Fiat","model":"500","color":"white"}',
+          'custom value',
+          'null',
+          'null',
+        ],
+      };
+      beforeEach(async () => {
+        create({
+          enhancedDatabaseReporting: true,
+        });
+      });
+
+      it('When enhancedDatabaseReporting:true, values should appear as parsable array of strings', done => {
+        const span = tracer.startSpan('test span');
+        context.with(trace.setSpan(context.active(), span), () => {
+          client.query(query, values, (err, res) => {
+            assert.strictEqual(err, null);
+            assert.ok(res);
+            runCallbackTest(span, attributes, events);
+            done();
+          });
+        });
       });
     });